Add files via upload
We created a couple of functions to correlate the name of the subcontractors we are looking for. We do two searches, the initial one searches for their names verbatim, and the second one searches for their names based on a desired permutation of the arrangement of their names, the latter removes the undesired words (such as "CO", "INC", etc.) and searches for the useful ones and returns a data frame with their names and emails. What's left to do is to connect the function with the google search function.